[yixuan-refactor] modularize chat_config
